Made sign up form its own component and added styled jsx
