+++ .babelrc
@@ -1,4 +1,4 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
-  "plugins": ["@babel/transform-runtime"]
+  "plugins": ["@babel/transform-runtime", "styled-jsx/babel"]
 }
