+++ client/pages/signup.jsx
@@ -1,144 +1,171 @@
-import React, { useState } from "react";
+import React from "react";
 import { ApolloConsumer } from "react-apollo";
-import gql from "graphql-tag";
+//import gql from "graphql-tag";
 
 import Layout from "../components/layouts/Layout";
+import SignUpBox from "../components/SignUpBox";
 
-export const addUserMutation = gql`
-  mutation addUser($email: String!, $password: String!) {
-    addUser(email: $email, password: $password) {
-      email
-      password
-    }
-  }
-`;
-
-export default () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [errorMessage, setErrorMessage] = useState("");
-  const [successMessage, setSuccessMessage] = useState("");
+const Signup = () => {
   return (
     <ApolloConsumer>
       {client => (
         <Layout>
-          <form className="form-signin mt-5">
-            <h1 className="mb-3">Sign Up</h1>
-            <div className="form-group">
-              <label htmlFor="email" className="float-left required">
-                Email address
-              </label>
-              <input
-                type="email"
-                className="form-control"
-                id="email"
-                placeholder="Enter your email"
-                value={email}
-                onChange={event => {
-                  setEmail(event.target.value);
-                }}
-                autoFocus
-                required
-              />
-              <small
-                id="emailHelp"
-                className="form-text text-muted float-left mb-3"
-              >
-                We&apos;ll never share your email with anyone else.
-              </small>
-              <br />
-            </div>
-            <div className="form-group">
-              <label htmlFor="password" className="float-left required">
-                Password
-              </label>
-              <input
-                type="password"
-                className="form-control"
-                id="password"
-                placeholder="Enter your password"
-                value={password}
-                onChange={event => {
-                  setPassword(event.target.value);
-                }}
-                required
-              />
-              <small
-                id="emailHelp"
-                className="form-text text-muted float-left mb-3"
-              >
-                Must contain at least 6 alphanumeric characters.
-              </small>
-              <br />
-            </div>
-
-            <button
-              type="submit"
-              className="btn btn-primary float-left"
-              onClick={async event => {
-                setErrorMessage("");
-                setSuccessMessage("");
-                event.preventDefault();
-                if (password.length < 6) {
-                  return setErrorMessage(
-                    "Password must be at least 6 characters in length"
-                  );
-                }
-                let passRegEx = /^[a-z0-9]+$/i;
-                if (passRegEx.test(password) === false) {
-                  return setErrorMessage(
-                    "Password must only contain alphanumeric characters"
-                  );
-                }
-
-                const { data } = await client.mutate({
-                  mutation: addUserMutation,
-                  variables: { email, password }
-                });
-                if (data.addUser.email === "empty") {
-                  setErrorMessage("You must enter a valid email address");
-                } else if (data.addUser.email === "taken") {
-                  setErrorMessage("That email address is already in use.");
-                } else {
-                  setSuccessMessage("Your account has been created!");
-                }
-                setEmail("");
-                setPassword("");
-              }}
-            >
-              Sign in
-            </button>
-          </form>
-          {errorMessage.length ? (
-            <p className="text-center text-danger">{errorMessage}</p>
-          ) : (
-            successMessage.length > 0 && (
-              <p className="text-center text-success">{successMessage}</p>
-            )
-          )}
-
-          <style>{`
-        .form-signin {
-          text-align: center;
-          width: 100%;
-          max-width: 400px;
-          margin: auto;
-          padding-top: 40px;
-          padding-bottom: 40px;
-        }
-
-        .form-signin .form-control {
-          padding: 10px;
-        }
-
-        .required:after{ 
-          content:'*'; 
-          color:red; 
-          padding-left:5px;
-        }
-      `}</style>
+          <div className="box-container">
+            <SignUpBox apolloClient={client} />
+          </div>
+          <style jsx>{`
+            .box-container {
+              text-align: center;
+              width: 100%;
+              max-width: 400px;
+              margin: auto;
+              padding-top: 40px;
+              padding-bottom: 40px;
+            }
+          `}</style>
         </Layout>
       )}
     </ApolloConsumer>
   );
 };
+
+export default Signup;
+
+// export const addUserMutation = gql`
+//   mutation addUser($email: String!, $password: String!) {
+//     addUser(email: $email, password: $password) {
+//       email
+//       password
+//     }
+//   }
+// `;
+
+// export default () => {
+//   const [email, setEmail] = useState("");
+//   const [password, setPassword] = useState("");
+//   const [errorMessage, setErrorMessage] = useState("");
+//   const [successMessage, setSuccessMessage] = useState("");
+//   return (
+//     <ApolloConsumer>
+//       {client => (
+//         <Layout>
+//           <form className="form-signin mt-5">
+//             <h1 className="mb-3">Sign Up</h1>
+//             <div className="form-group">
+//               <label htmlFor="email" className="float-left required">
+//                 Email address
+//               </label>
+//               <input
+//                 type="email"
+//                 className="form-control"
+//                 id="email"
+//                 placeholder="Enter your email"
+//                 value={email}
+//                 onChange={event => {
+//                   setEmail(event.target.value);
+//                 }}
+//                 autoFocus
+//                 required
+//               />
+//               <small
+//                 id="emailHelp"
+//                 className="form-text text-muted float-left mb-3"
+//               >
+//                 We&apos;ll never share your email with anyone else.
+//               </small>
+//               <br />
+//             </div>
+//             <div className="form-group">
+//               <label htmlFor="password" className="float-left required">
+//                 Password
+//               </label>
+//               <input
+//                 type="password"
+//                 className="form-control"
+//                 id="password"
+//                 placeholder="Enter your password"
+//                 value={password}
+//                 onChange={event => {
+//                   setPassword(event.target.value);
+//                 }}
+//                 required
+//               />
+//               <small
+//                 id="emailHelp"
+//                 className="form-text text-muted float-left mb-3"
+//               >
+//                 Must contain at least 6 alphanumeric characters.
+//               </small>
+//               <br />
+//             </div>
+
+//             <button
+//               type="submit"
+//               className="btn btn-primary float-left"
+//               onClick={async event => {
+//                 setErrorMessage("");
+//                 setSuccessMessage("");
+//                 event.preventDefault();
+//                 if (password.length < 6) {
+//                   return setErrorMessage(
+//                     "Password must be at least 6 characters in length"
+//                   );
+//                 }
+//                 let passRegEx = /^[a-z0-9]+$/i;
+//                 if (passRegEx.test(password) === false) {
+//                   return setErrorMessage(
+//                     "Password must only contain alphanumeric characters"
+//                   );
+//                 }
+
+//                 const { data } = await client.mutate({
+//                   mutation: addUserMutation,
+//                   variables: { email, password }
+//                 });
+//                 if (data.addUser.email === "empty") {
+//                   setErrorMessage("You must enter a valid email address");
+//                 } else if (data.addUser.email === "taken") {
+//                   setErrorMessage("That email address is already in use.");
+//                 } else {
+//                   setSuccessMessage("Your account has been created!");
+//                 }
+//                 setEmail("");
+//                 setPassword("");
+//               }}
+//             >
+//               Sign in
+//             </button>
+//           </form>
+//           {errorMessage.length ? (
+//             <p className="text-center text-danger">{errorMessage}</p>
+//           ) : (
+//             successMessage.length > 0 && (
+//               <p className="text-center text-success">{successMessage}</p>
+//             )
+//           )}
+
+//           <style>{`
+//         .form-signin {
+//           text-align: center;
+//           width: 100%;
+//           max-width: 400px;
+//           margin: auto;
+//           padding-top: 40px;
+//           padding-bottom: 40px;
+//         }
+
+//         .form-signin .form-control {
+//           padding: 10px;
+//         }
+
+//         .required:after{
+//           content:'*';
+//           color:red;
+//           padding-left:5px;
+//         }
+//       `}</style>
+//         </Layout>
+//       )}
+//     </ApolloConsumer>
+//   );
+// };
