+++ package.json
@@ -44,7 +44,8 @@
     "pino": "^5.11.3",
     "react": "^16.8.4",
     "react-apollo": "^2.5.2",
-    "react-dom": "^16.8.4"
+    "react-dom": "^16.8.4",
+    "styled-jsx": "^3.2.1"
   },
   "devDependencies": {
     "@babel/node": "^7.2.2",
